[slicer] Slicer now works for base case
